requirements.txt
google-auth==2.28.1
google-cloud-resource-manager==1.12.3
google-api-python-client==2.121.0
google-cloud-billing==1.13.3
google-cloud-service-usage==1.10.3
google-cloud-iam==2.14.3
googleapis-common-protos==1.62.0
python-dotenv==1.0.1
google-cloud-compute